Extract Class를 사용하여 한 파일 내에 있던 Class들을 패키지 레벨로 격상
