+++ src/main/kotlin/com/example/ad/contract/ErrorMessage.kt
@@ -1,0 +1,6 @@
+package com.example.ad.contract
+
+enum class ErrorMessage(val value: String) {
+    UNKNOWN_INPUT("처리 할 수 없는 입력 값입니다"),
+    NOT_VALID_INPUT("입력 값이 유효하지 않습니다"),
+}
