+++ src/main/kotlin/com/example/ad/contract/InvalidCampaignTypeException.kt
@@ -1,0 +1,3 @@
+package com.example.ad.contract
+
+class InvalidCampaignTypeException : ValidationException(ValidationErrors.invalidCampaignType())
