+++ src/main/kotlin/com/example/ad/contract/ValidationError.kt
@@ -1,0 +1,6 @@
+package com.example.ad.contract
+
+data class ValidationError(
+    val field: String,
+    val reason: String?,
+)
