+++ src/main/kotlin/com/example/ad/contract/ValidationException.kt
@@ -1,0 +1,3 @@
+package com.example.ad.contract
+
+sealed class ValidationException(val validationError: ValidationError) : RuntimeException()
